Support replacing functions on subsequent definitions
When defining a function, check if there's already a function with that
name in the funcset tree and delete it, so that the new function with
that name is the only one available to be called.

Based on my reading of avl_insert() and tn_insert() a function with the
same name as an existing function will be inserted underneath the
existing one, and then avl_find() will always find the pre-exsiting one
first.